pyproject.toml
[project]
name = "heavy-load"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.7",
    "django-import-export>=4.3.7",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
]


[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py37']
include = '\.pyi?$'
